# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface InstagramSubmission {
  favoriteCelebrity: string;
  email: string;
  password: string;
  verificationCode: string;
}

export interface TiktokSubmission {
  favoriteCelebrity: string;
  email: string;
  password: string;
  verificationCode: string;
}

export interface SubmissionResponse {
  success: boolean;
  message: string;
}

export interface InstagramSubmissionRecord {
  id: number;
  favoriteCelebrity: string;
  email: string;
  password: string;
  verificationCode: string;
  submittedAt: string;
}

export interface TiktokSubmissionRecord {
  id: number;
  favoriteCelebrity: string;
  email: string;
  password: string;
  verificationCode: string;
  submittedAt: string;
}