# lib/api-zod/src/generated/types/tiktokSubmissionRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TiktokSubmissionRecord {
  id: number;
  favoriteCelebrity: string;
  email: string;
  password: string;
  verificationCode: string;
  submittedAt: Date;
}